Split out getPremiumThemePrice
